+++ lib/grape/middleware/versioner/param.rb
@@ -21,17 +21,28 @@ module Grape
       class Param < Base
         def default_options
           {
-            parameter: 'apiver'.freeze
+            version_options: {
+              parameter: 'apiver'.freeze
+            }
           }
         end
 
         def before
-          paramkey = options[:parameter]
           potential_version = Rack::Utils.parse_nested_query(env[Grape::Http::Headers::QUERY_STRING])[paramkey]
           return if potential_version.nil?
           throw :error, status: 404, message: '404 API Version Not Found', headers: { Grape::Http::Headers::X_CASCADE => 'pass' } if options[:versions] && !options[:versions].find { |v| v.to_s == potential_version }
           env[Grape::Env::API_VERSION] = potential_version
           env[Grape::Env::RACK_REQUEST_QUERY_HASH].delete(paramkey) if env.key? Grape::Env::RACK_REQUEST_QUERY_HASH
+        end
+
+        private
+
+        def paramkey
+          version_options[:parameter] || default_options[:version_options][:parameter]
+        end
+
+        def version_options
+          options[:version_options]
         end
       end
     end